finally, we can move fighters on server
TODO: draw scene and fighters movement on client

+++ api/application/db/DB.php
@@ -29,7 +29,8 @@ class DB {
         }
         return $res;
     }
-    //users and auth
+
+    /* USER */
     public function registerUser($login, $pass, $token) {
         $query = "INSERT INTO users (login, password, token) VALUES ( '".$login."', '".$pass."', '".$token."')";
         $result = $this->connection->query($query);
@@ -41,18 +42,39 @@ class DB {
         $result = $this->connection->query($query);
         return $this->oneRecord($result); 
     }
-    public function updateUserToken($id, $token) {
-        $query = "UPDATE users SET token = '".$token."' WHERE id = ".$id."";
-        $result = $this->connection->query($query);
-        return true;
-    }
-
+    
     public function getUserByToken($token) {
         $query = "SELECT * FROM users WHERE token = '".$token."'";
         $result = $this->connection->query($query);
         return $this->oneRecord($result);
     }
-    //lobby
+
+    public function updateUserToken($id, $token) {
+        $query = "UPDATE users SET token = '".$token."' WHERE id = '".$id."'";
+        $result = $this->connection->query($query);
+        return true;
+    }
+    /* LOBBY */
+    public function newChallenge($userId1, $userId2) {
+        $query = "INSERT INTO lobby (id_user1, id_user2, status) VALUES ('".$userId1."', '".$userId2."', 'open')";
+        $result = $this->connection->query($query);
+        return true;
+    }
+
+    public function isChallenge($userId) {
+        $query = "SELECT * 
+                  FROM lobby 
+                  WHERE id_user2 = ".$userId." AND status = 'open' OR id_user1 = '".$userId."' AND status = 'open'";
+        $result = $this->connection->query($query);
+        return $this->oneRecord($result);
+    }
+
+    public function isChallengeAccepted($userId) {
+        $query = "SELECT * FROM lobby WHERE id_user1 = '".$userId."' AND status = 'game'";
+        $result = $this->connection->query($query);
+        return $this->oneRecord($result);
+    }
+    
     public function getLobbyUsers($userId) {
         $query = "SELECT id, login
                 FROM users
@@ -68,55 +90,35 @@ class DB {
                             b.status='open' AND
                             (b.id_fighter1 = f.id OR 
                              b.id_fighter2 = f.id)) AND
-                    id <> ".$userId."";
+                    id <> '".$userId."'";
         $result = $this->connection->query($query);
         return $this->allRecords($result);
     }
-
-    public function deleteOldUserChallenge($userId) {
-        $query = "DELETE FROM lobby WHERE id_user1 = ".$userId."";
-        $result = $this->connection->query($query);
-        return true;
-    }
-
-    public function newChallenge($userId1, $userId2) {
-        $query = "INSERT INTO lobby (id_user1, id_user2, status) VALUES (".$userId1.", ".$userId2.", 'open')";
-        $result = $this->connection->query($query);
-        return true;
-    }
-
-    public function isChallenge($userId) {
-        $query = "SELECT * 
-                  FROM lobby 
-                  WHERE id_user2 = ".$userId." AND status = 'open' OR id_user1 = ".$userId." AND status = 'open'";
-        $result = $this->connection->query($query);
-        return $this->oneRecord($result);
-    }
-
-    public function isChallengeAccepted($userId) {
-        $query = "SELECT * FROM lobby WHERE id_user1 = ".$userId." AND status = 'game'";
-        $result = $this->connection->query($query);
-        return $this->oneRecord($result);
-    }
-
-    public function acceptChallenge($userId2, $answer) {
-        $query = "UPDATE lobby SET status = '".$answer."' WHERE id_user2 = ".$userId2."";
-        $result = $this->connection->query($query);
-        return true;
-    }
-
+    
     public function getLobbyInGame($userId2){
-        $query = "SELECT * FROM lobby WHERE id_user2 = ".$userId2." AND status = 'game'";
+        $query = "SELECT * FROM lobby WHERE id_user2 = '".$userId2."' AND status = 'game'";
         $result = $this->connection->query($query);
         return $this->oneRecord($result); 
     }
-    //game
-    public function deleteFighterByUserId($userId) {
-        $query = "DELETE FROM fighters WHERE user_id = ".$userId."";
+
+    public function acceptChallenge($userId2, $answer) {
+        $query = "UPDATE lobby SET status = '".$answer."' WHERE id_user2 = '".$userId2."'";
         $result = $this->connection->query($query);
         return true;
     }
 
+    public function deleteOldUserChallenge($userId) {
+        $query = "DELETE FROM lobby WHERE id_user1 = '".$userId."'";
+        $result = $this->connection->query($query);
+        return true;
+    }
+
+    public function deleteLobby($userId) {
+        $query = "DELETE FROM lobby WHERE id_user1 = '".$userId."' OR id_user2 = '".$userId."'";
+        $result = $this->connection->query($query);
+        return true;
+    }
+    /* FIGHTER */
     public function createFighter($data) {
         $query = "INSERT INTO fighters 
                   (user_id, x, y, width, height, state, direction, health)
@@ -132,17 +134,42 @@ class DB {
         return true;
     }
 
+    public function getFighter($fighterId) {
+        $query = "SELECT * FROM fighters WHERE id = ".$fighterId."";
+        $result = $this->connection->query($query);
+        return $this->oneRecord($result);
+    }
+
     public function getFighterByUserId($userId) {
         $query = "SELECT * FROM fighters WHERE user_id = ".$userId."";
         $result = $this->connection->query($query);
         return $this->oneRecord($result);
     }
 
-    public function createBattle($fighterId1, $fighterId2) {
+    public function moveFighter($fighterId, $x, $direction) {
+        $query = "UPDATE fighters SET x = '".$x."', direction = '".$direction."' WHERE id = '".$fighterId."'";
+        $result = $this->connection->query($query);
+        return true;
+    }
+
+    public function deleteFighterById($fighterId) {
+        $query = "DELETE FROM fighters WHERE id = ".$fighterId."";
+        $result = $this->connection->query($query);
+        return true;
+    }
+
+    public function deleteFighterByUserId($userId) {
+        $query = "DELETE FROM fighters WHERE user_id = ".$userId."";
+        $result = $this->connection->query($query);
+        return true;
+    }
+    /* BATTLE */
+    public function createBattle($fighterId1, $fighterId2, $startTimestamp) {
         $query = "INSERT INTO battles
-                  (id_fighter1, id_fighter2) 
+                  (id_fighter1, id_fighter2, startTimestamp ) 
                   VALUES (".$fighterId1.",
-                         ".$fighterId2.")";
+                         ".$fighterId2.",
+                         ".$startTimestamp.")";
         $result = $this->connection->query($query);
         return true;
     }
@@ -153,12 +180,12 @@ class DB {
         return $this->oneRecord($result);
     }
 
-    public function getFighter($fighterId) {
-        $query = "SELECT * FROM fighters WHERE id = ".$fighterId."";
+    public function updateBattleTimestamp($battleId, $newTimestamp) {
+        $query = "UPDATE battles SET timestamp = '".$newTimestamp."' WHERE id = '".$battleId."'";
         $result = $this->connection->query($query);
-        return $this->oneRecord($result);
+        return true;
     }
-   
+
     public function deleteBattle($battleId) {
         $query = "DELETE FROM battles WHERE id = ".$battleId."";
         $result = $this->connection->query($query);
@@ -173,22 +200,4 @@ class DB {
         return true;
     }
 
-    public function deleteLobby($userId) {
-        $query = "DELETE FROM lobby WHERE id_user1 = $userId OR id_user2 = $userId";
-        $result = $this->connection->query($query);
-        return true;
-    }
-
-    public function updateBattleTimestamp($battleId, $newTimestamp) {
-        $query = "UPDATE battles SET timestamp = $newTimestamp WHERE id = $battleId";
-        $result = $this->connection->query($query);
-        return true;
-    }
-
-    public function deleteFighterById($fighterId) {
-        $query = "DELETE FROM fighters WHERE id = ".$fighterId."";
-        $result = $this->connection->query($query);
-        return true;
-    }
-    
 }
